Create SonyCameraRemoteControl.ino
Connects to a Sony cameras WiFi when running the Remote Camera Control APP.
It then sends a message to the camera to take a picture when either the button is pressed or the second ESP pulls its trigger input low.

Most of this code was written by Reinhard Nickels. I only modifiesd it to work for my application of taking timelapses with a Bambu Lab 3D printer.